finished bonus, debugged compilation error, now need to protect

+++ calloc.c
@@ -6,7 +6,7 @@
 /*   By: rponsonn <rponsonn@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/15 14:05:04 by rponsonn          #+#    #+#             */
-/*   Updated: 2021/01/15 14:25:55 by rponsonn         ###   ########.fr       */
+/*   Updated: 2021/01/24 16:48:31 by rponsonn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_calloc(size_t count, size_t size)
 {
 	char	*ptr;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	if (!(ptr = malloc((size) * count)))
